Remove reliance of QuickJS on fenv.h.
This puts calls to fenv functions behind an if/def which is not set.
I believe this makes QuickJS not fully standards compliant with
regards to numerical rounding, especially when converting floating
point numbers to strings. I don't think PuzzleScript relies on
any precise rounding when converting numbers to strings so this
should be safe for PuzzleScript.